refactor(pack): extract business logic to service layer for pack controller
