+++ server/routes/pack/pack-controller.ts
@@ -1,7 +1,5 @@
 import db from '../../db/connection.js';
 import { Tables } from '../../db/tables.js';
-import { DEFAULT_PALETTE_COLOR } from '../../utils/constants.js';
-import { validateWeightUnit, validatePrice } from './pack-utils.js';
 import { Request, Response } from 'express';
 import { MulterS3File } from '../../types/multer.js';
 import {
@@ -14,15 +12,7 @@ import {
 	ErrorCode,
 } from '../../utils/error-response.js';
 import { createCloudfrontUrlForPhoto, s3DeletePhoto } from '../../utils/s3.js';
-import { packScraper } from './pack-scraper.js';
 import { isError } from '../../utils/validation.js';
-import {
-	DEFAULT_INCREMENT,
-	getNextAppendIndex,
-	moveWithFractionalIndex,
-	bulkMoveToGearCloset,
-	type MovablePackItem,
-} from '../../utils/fractional-indexing/fractional-indexing.js';
 import { logError, logger } from '../../config/logger.js';
 import {
 	hasEmptyValidatedBody,
@@ -41,8 +31,24 @@ import {
 	PackCategoryMove,
 	packFields,
 	packItemFields,
-	packCategoryFields,
 } from './pack-schemas.js';
+
+import {
+	getPackWithCategories,
+	createNewPack,
+	importPackFromUrl,
+	movePackInList,
+	deletePackAndMoveItems,
+	deletePackCompletely,
+	createPackItem,
+	movePackItemBetweenCategories,
+	moveItemToGearCloset,
+	createPackCategory,
+	movePackCategory as movePackCategoryService,
+	moveCategoryItemsToCloset,
+	deleteCategoryAndAllItems,
+	getAvailablePacks as getAvailablePacksService,
+} from './pack-service.js';
 
 async function getDefaultPack(req: Request, res: Response) {
 	try {
@@ -55,30 +61,7 @@ async function getDefaultPack(req: Request, res: Response) {
 			.first();
 
 		if (pack) {
-			// Gets categories for the pack ordered by category index
-			// Groups all pack items for each category ordered by pack item index
-			// Build pack item JSON object for aggregation
-			const packItemJson = packItemFields.map((f) => `'${f}', pi.${f}`).join(', ');
-
-			const result = await db.raw(
-				`select 
-					pc.*, 
-					COALESCE(
-						json_agg(
-							json_build_object(${packItemJson})
-							ORDER BY pi.pack_item_index::NUMERIC
-						) FILTER (WHERE pi.pack_item_id IS NOT NULL),
-						'[]'::json
-					) as pack_items 
-					from pack_category pc
-					left outer join pack_item pi on pi.pack_category_id = pc.pack_category_id	
-				where pc.user_id = ? and pc.pack_id = ?
-				group by pc.pack_category_id
-				order by pc.pack_category_index::NUMERIC`,
-				[userId, pack.pack_id],
-			);
-			const { rows: categories = [] } = result;
-
+			const { categories } = await getPackWithCategories(userId, pack.pack_id!);
 			return successResponse(res, { pack, categories });
 		} else {
 			// UI can handle not having a default pack.
@@ -95,7 +78,7 @@ async function getPack(req: Request, res: Response) {
 		const { userId } = req;
 		const { packId } = req.params;
 
-		const { pack, categories } = await getPackById(userId, Number(packId));
+		const { pack, categories } = await getPackWithCategories(userId, Number(packId));
 
 		if (!pack) {
 			return notFound(res, 'Pack not found.');
@@ -111,45 +94,10 @@ async function getPack(req: Request, res: Response) {
 	}
 }
 
-async function getPackById(userId: string, packId: number) {
-	const pack = await db(Tables.Pack)
-		.select(packFields)
-		.where('user_id', userId)
-		.where('pack_id', packId)
-		.orderBy('created_at')
-		.first();
-
-	// Gets categories for a pack ordered by category index
-	// Groups all pack items for each category ordered by pack item index
-	// Build pack item JSON object for aggregation
-	const packItemJson = packItemFields.map((f) => `'${f}', pi.${f}`).join(', ');
-
-	const result = await db.raw(
-		`select 
-			pc.*, 
-			COALESCE(
-				json_agg(
-					json_build_object(${packItemJson})
-					ORDER BY pi.pack_item_index::NUMERIC
-				) FILTER (WHERE pi.pack_item_id IS NOT NULL),
-				'[]'::json
-			) as pack_items 
-			from pack_category pc
-			left outer join pack_item pi on pi.pack_category_id = pc.pack_category_id	
-		where pc.user_id = ? and pc.pack_id = ?
-		group by pc.pack_category_id
-		order by pc.pack_category_index::NUMERIC`,
-		[userId, packId],
-	);
-	const { rows: categories = [] } = result;
-
-	return { pack: pack || null, categories };
-}
-
 async function getPackList(req: Request, res: Response) {
 	try {
 		const { userId } = req;
-		const packList = await getAvailablePacks(userId);
+		const packList = await getAvailablePacksService(userId);
 
 		return successResponse(res, { packList });
 	} catch (err) {
@@ -183,142 +131,13 @@ async function addNewPack(req: Request, res: Response) {
 	}
 }
 
-async function createNewPack(userId: string) {
-	// Calculate index for new pack (append to end of user's pack list)
-	const packIndex = await getNextAppendIndex(Tables.Pack, 'pack_index', {
-		user_id: userId,
-	});
-
-	const trx = await db.transaction();
-	try {
-		const [pack] = await trx(Tables.Pack)
-			.insert({
-				user_id: userId,
-				pack_name: 'New Pack',
-				pack_index: packIndex,
-			})
-			.returning(packFields);
-
-		const categories = await trx(Tables.PackCategory)
-			.insert({
-				user_id: userId,
-				pack_id: pack.pack_id,
-				pack_category_name: '',
-				pack_category_index: DEFAULT_INCREMENT.toString(),
-				pack_category_color: DEFAULT_PALETTE_COLOR,
-			})
-			.returning(packCategoryFields);
-
-		const packItems = await trx(Tables.PackItem)
-			.insert({
-				user_id: userId,
-				pack_id: pack.pack_id,
-				pack_category_id: categories[0].pack_category_id,
-				pack_item_name: '',
-				pack_item_index: DEFAULT_INCREMENT.toString(),
-			})
-			.returning(packItemFields);
-
-		const categoriesWithItems = categories.map((cat, index) => ({
-			...cat,
-			pack_items: index === 0 ? packItems : [],
-		}));
-
-		await trx.commit();
-		return { pack, categories: categoriesWithItems };
-	} catch (err) {
-		await trx.rollback();
-		throw new Error('There was an error creating a new pack.');
-	}
-}
-
 async function importNewPack(req: ValidatedRequest<PackMigration>, res: Response) {
 	const importErrorMessage = 'There was an error importing your pack.';
 	try {
 		const { userId } = req;
 		const { pack_url, palette_list } = req.validatedBody;
-		const importedPack = await packScraper(pack_url);
 
-		// handle error
-		const isPackError = isError(importedPack);
-		if (isPackError) {
-			logger.warn('Pack import failed - invalid pack data', {
-				userId,
-				packUrl: pack_url,
-				error: 'Pack scraper returned error or invalid data',
-				scrapedData: importedPack,
-			});
-			return badRequest(res, importErrorMessage);
-		}
-
-		// save new pack to db
-		const { pack_name, pack_description, pack_categories } = importedPack;
-
-		// Calculate index for new imported pack (append to end of user's pack list)
-		const packIndex = await getNextAppendIndex(Tables.Pack, 'pack_index', {
-			user_id: userId,
-		});
-
-		const trx = await db.transaction();
-		try {
-			// insert pack
-			const [{ pack_id }] = await trx(Tables.Pack)
-				.insert({
-					user_id: userId,
-					pack_name,
-					pack_description,
-					pack_index: packIndex,
-				})
-				.returning('pack_id');
-
-			// insert categories and pack items sequentially
-			for (const category of pack_categories) {
-				const { pack_category_name, pack_category_index, pack_items } = category;
-
-				// get theme color based on index
-				const themeColor = palette_list?.[pack_category_index % palette_list.length];
-
-				// assign provided palette or default
-				const pack_category_color = isError(themeColor)
-					? DEFAULT_PALETTE_COLOR
-					: themeColor;
-
-				// insert pack category
-				const [{ pack_category_id }] = await trx(Tables.PackCategory)
-					.insert({
-						user_id: userId,
-						pack_id,
-						pack_category_name,
-						pack_category_index: pack_category_index.toString(),
-						pack_category_color,
-					})
-					.returning('pack_category_id');
-
-				// insert pack items
-				if (pack_items.length > 0) {
-					const packItemsWithIds = pack_items.map((item) => ({
-						...item,
-						user_id: userId,
-						pack_id,
-						pack_category_id,
-						pack_item_index: item.pack_item_index?.toString(),
-						pack_item_weight_unit: validateWeightUnit(item.pack_item_weight_unit),
-						pack_item_price: validatePrice(item.pack_item_price),
-					}));
-					await trx(Tables.PackItem).insert(packItemsWithIds);
-				}
-			}
-
-			await trx.commit();
-		} catch (error) {
-			await trx.rollback();
-			throw new Error('Failed to import pack');
-		}
-
-		logger.info('User imported external pack successfully', {
-			userId,
-			packUrl: pack_url,
-		});
+		await importPackFromUrl(userId, pack_url, palette_list);
 
 		return successResponse(res, null, 'Pack imported successfully');
 	} catch (err) {
@@ -458,25 +277,9 @@ async function movePack(req: ValidatedRequest<PackMove>, res: Response) {
 		const { packId } = req.params;
 		const { prev_pack_index, next_pack_index } = req.validatedBody;
 
-		const { newIndex, rebalanced } = await moveWithFractionalIndex(
-			Tables.Pack,
-			'pack_index',
-			'pack_id',
-			packId,
-			prev_pack_index,
-			next_pack_index,
-			{ user_id: userId }, // WHERE conditions
-		);
+		const result = await movePackInList(userId, packId, prev_pack_index, next_pack_index);
 
-		return successResponse(
-			res,
-			{
-				packId,
-				newIndex,
-				rebalanced,
-			},
-			'Pack moved successfully',
-		);
+		return successResponse(res, result, 'Pack moved successfully');
 	} catch (err) {
 		return internalError(res, 'There was an error changing your pack order.');
 	}
@@ -487,33 +290,9 @@ async function deletePack(req: Request, res: Response) {
 		const { userId } = req;
 		const { packId } = req.params;
 
-		// Get all items from the pack to move to gear closet
-		const packItems = await db(Tables.PackItem)
-			.select(`${Tables.PackItem}.*`)
-			.join(
-				Tables.PackCategory,
-				`${Tables.PackItem}.pack_category_id`,
-				`${Tables.PackCategory}.pack_category_id`,
-			)
-			.where(`${Tables.PackItem}.user_id`, userId)
-			.where(`${Tables.PackItem}.pack_id`, packId)
-			.orderByRaw(
-				`${Tables.PackCategory}.pack_category_index::NUMERIC, ${Tables.PackItem}.pack_item_index::NUMERIC`,
-			);
+		const result = await deletePackAndMoveItems(userId, packId);
 
-		await bulkMoveToGearCloset(packItems, userId);
-
-		// DELETE ON CASCADE: pack -> pack_category -> pack_item
-		await db(Tables.Pack).del().where('user_id', userId).where('pack_id', packId);
-
-		//if no packs left, create default pack
-		const response = await db(Tables.Pack)
-			.select('pack_id')
-			.where('user_id', userId)
-			.first();
-		if (!response) await createNewPack(userId);
-
-		return successResponse(res, { deletedPackId: packId }, 'Pack deleted successfully');
+		return successResponse(res, result, 'Pack deleted successfully');
 	} catch (err) {
 		return internalError(res, 'There was an error deleting your pack.');
 	}
@@ -524,21 +303,9 @@ async function deletePackAndItems(req: Request, res: Response) {
 		const { userId } = req;
 		const { packId } = req.params;
 
-		// DELETE ON CASCADE: pack -> pack_category -> pack_item
-		await db(Tables.Pack).del().where('user_id', userId).where('pack_id', packId);
+		const result = await deletePackCompletely(userId, packId);
 
-		// if no packs left, ensure user has default pack
-		const response = await db(Tables.Pack)
-			.select('pack_id')
-			.where('user_id', userId)
-			.first();
-		if (!response) await createNewPack(userId);
-
-		return successResponse(
-			res,
-			{ deletedPackId: packId },
-			'Pack and items deleted successfully',
-		);
+		return successResponse(res, result, 'Pack and items deleted successfully');
 	} catch (err) {
 		return internalError(res, 'There was an error deleting your pack.');
 	}
@@ -549,25 +316,9 @@ async function addPackItem(req: ValidatedRequest<PackItemCreate>, res: Response)
 		const { userId } = req;
 		const { pack_id, pack_category_id } = req.validatedBody;
 
-		// Calculate index for new item
-		const pack_item_index = await getNextAppendIndex(Tables.PackItem, 'pack_item_index', {
-			user_id: userId,
-			pack_id,
-			pack_category_id,
-		});
+		const result = await createPackItem(userId, pack_id, pack_category_id);
 
-		const [packItem] =
-			(await db(Tables.PackItem)
-				.insert({
-					user_id: userId,
-					pack_id,
-					pack_category_id,
-					pack_item_name: '',
-					pack_item_index,
-				})
-				.returning(packItemFields)) || [];
-
-		return successResponse(res, { packItem }, 'Pack item added successfully');
+		return successResponse(res, result, 'Pack item added successfully');
 	} catch (err) {
 		return internalError(res, 'There was an error adding a new pack item.');
 	}
@@ -602,26 +353,16 @@ async function movePackItem(req: ValidatedRequest<PackItemMove>, res: Response) 
 		const { pack_category_id, prev_pack_category_id, prev_item_index, next_item_index } =
 			req.validatedBody;
 
-		const { newIndex, rebalanced } = await moveWithFractionalIndex(
-			Tables.PackItem,
-			'pack_item_index',
-			'pack_item_id',
+		const result = await movePackItemBetweenCategories(
+			userId,
 			packItemId,
+			pack_category_id,
+			prev_pack_category_id,
 			prev_item_index,
 			next_item_index,
-			{ user_id: userId, pack_category_id: prev_pack_category_id }, // WHERE conditions - use PREVIOUS category to find the item
-			{ pack_category_id }, // Additional fields to update (can change category on drag/drop)
 		);
 
-		return successResponse(
-			res,
-			{
-				newIndex,
-				rebalanced,
-				categoryChanged: prev_pack_category_id !== pack_category_id,
-			},
-			'Pack item moved successfully',
-		);
+		return successResponse(res, result, 'Pack item moved successfully');
 	} catch (err) {
 		logError('Move pack item failed', err, {
 			userId: req.userId,
@@ -636,25 +377,7 @@ async function moveItemToCloset(req: Request, res: Response) {
 		const { userId } = req;
 		const { packItemId } = req.params;
 
-		// Calculate index for item moving to gear closet
-		const newGearClosetIndex = await getNextAppendIndex(
-			Tables.PackItem,
-			'pack_item_index',
-			{
-				user_id: userId,
-				pack_id: null,
-				pack_category_id: null,
-			},
-		);
-
-		await db(Tables.PackItem)
-			.update({
-				pack_id: null,
-				pack_category_id: null,
-				pack_item_index: newGearClosetIndex,
-			})
-			.where('user_id', userId)
-			.where('pack_item_id', packItemId);
+		await moveItemToGearCloset(userId, packItemId);
 
 		return successResponse(res, null, 'Item moved to gear closet successfully');
 	} catch (err) {
@@ -684,44 +407,9 @@ async function addPackCategory(req: ValidatedRequest<PackCategoryCreate>, res: R
 		const { packId } = req.params;
 		const { category_color } = req.validatedBody;
 
-		// Calculate index for new category
-		const packCategoryIndex = await getNextAppendIndex(
-			Tables.PackCategory,
-			'pack_category_index',
-			{ user_id: userId, pack_id: Number(packId) },
-		);
+		const result = await createPackCategory(userId, packId, category_color);
 
-		const [packCategory] = await db(Tables.PackCategory)
-			.insert({
-				pack_category_name: '',
-				user_id: userId,
-				pack_id: Number(packId),
-				pack_category_index: packCategoryIndex.toString(),
-				pack_category_color: category_color,
-			})
-			.returning(packCategoryFields);
-
-		// add default pack item (can be default index for first item)
-		const [packItem] = await db(Tables.PackItem)
-			.insert({
-				user_id: userId,
-				pack_id: Number(packId),
-				pack_category_id: packCategory.pack_category_id,
-				pack_item_name: '',
-				pack_item_index: DEFAULT_INCREMENT.toString(),
-			})
-			.returning(packItemFields);
-
-		const categoryWithItems = {
-			...packCategory,
-			pack_items: [packItem],
-		};
-
-		return successResponse(
-			res,
-			{ packCategory: categoryWithItems },
-			'Pack category added successfully',
-		);
+		return successResponse(res, result, 'Pack category added successfully');
 	} catch (err) {
 		return internalError(res, 'There was an error adding a new category.');
 	}
@@ -756,24 +444,14 @@ async function movePackCategory(req: ValidatedRequest<PackCategoryMove>, res: Re
 		const { categoryId } = req.params;
 		const { prev_category_index, next_category_index } = req.validatedBody;
 
-		const { newIndex, rebalanced } = await moveWithFractionalIndex(
-			Tables.PackCategory,
-			'pack_category_index',
-			'pack_category_id',
+		const result = await movePackCategoryService(
+			userId,
 			categoryId,
 			prev_category_index,
 			next_category_index,
-			{ user_id: userId }, // WHERE conditions
 		);
 
-		return successResponse(
-			res,
-			{
-				newIndex,
-				rebalanced,
-			},
-			'Pack category moved successfully',
-		);
+		return successResponse(res, result, 'Pack category moved successfully');
 	} catch (err) {
 		logError('Move pack category failed', err, {
 			userId: req.userId,
@@ -788,23 +466,9 @@ async function moveCategoryToCloset(req: Request, res: Response) {
 		const { userId } = req;
 		const { categoryId } = req.params;
 
-		// Get all items from the category to move to gear closet
-		const categoryItems = await db(Tables.PackItem)
-			.select('*')
-			.where('user_id', userId)
-			.where('pack_category_id', categoryId)
-			.orderByRaw('pack_item_index::NUMERIC');
+		const result = await moveCategoryItemsToCloset(userId, categoryId);
 
-		// Bulk move all category items to gear closet (indexes will be reassigned)
-		await bulkMoveToGearCloset(categoryItems as MovablePackItem[], userId);
-
-		await deleteCategory(userId, categoryId);
-
-		return successResponse(
-			res,
-			{ deletedId: categoryId },
-			'Category moved to gear closet successfully',
-		);
+		return successResponse(res, result, 'Category moved to gear closet successfully');
 	} catch (err) {
 		return internalError(res, 'There was an error deleting your category.');
 	}
@@ -815,59 +479,11 @@ async function deleteCategoryAndItems(req: Request, res: Response) {
 		const { userId } = req;
 		const { categoryId } = req.params;
 
-		const trx = await db.transaction();
-		try {
-			await trx(Tables.PackItem)
-				.del()
-				.where('user_id', userId)
-				.where('pack_category_id', categoryId);
-			await trx(Tables.PackCategory)
-				.del()
-				.where('user_id', userId)
-				.where('pack_category_id', categoryId);
-			await trx.commit();
-		} catch (error) {
-			await trx.rollback();
-			throw new Error('Failed to delete category and items');
-		}
+		const result = await deleteCategoryAndAllItems(userId, categoryId);
 
-		return successResponse(
-			res,
-			{ deletedId: categoryId },
-			'Category and items deleted successfully',
-		);
+		return successResponse(res, result, 'Category and items deleted successfully');
 	} catch (err) {
 		return internalError(res, 'There was an error deleting your pack items.');
-	}
-}
-
-async function getAvailablePacks(userId: string) {
-	try {
-		const { rows = [] } = await db.raw(
-			`select 
-			pack.pack_id, pack_name, pack_index,
-			coalesce(array_remove(array_agg(to_jsonb(pc) order by pc.pack_category_index::NUMERIC), NULL), '{}') as pack_categories from pack
-			left outer join pack_category pc on pc.pack_id = pack.pack_id	
-		where pack.user_id = ?
-		group by pack.pack_id
-		order by pack.pack_index::NUMERIC`,
-			[userId],
-		);
-		return rows;
-	} catch (err) {
-		return new Error('Error getting available packs.');
-	}
-}
-
-async function deleteCategory(user_id: string, pack_category_id: number | string) {
-	try {
-		// Delete ON CASCADE: pack_category -> pack_item
-		await db(Tables.PackCategory)
-			.del()
-			.where('user_id', user_id)
-			.where('pack_category_id', pack_category_id);
-	} catch (err) {
-		throw new Error('Failed to delete category');
 	}
 }
